Add a basic landing page for when nothing is configured
